Split out test_date
